pyproject.toml
[build-system]
requires = [
    "setuptools>=68",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "ai-pi"
version = "0.0.0"
description = "ai-pi"
readme = "README.md"
authors = [
    { name = "Christian D'Andrea", email = "crdandre@gmail.com" },
]
dependencies = [
    "bayoo-docx ~= 0.2.20",
    "dspy ~= 2.5.32",
    "einops ~= 0.8.0",
    "fuzzywuzzy ~= 0.18.0",
    "python-dotenv ~= 1.0.1",
    "llama-index-core",
    "llama-index-readers-file",
    "llama-index-llms-nvidia", 
    "llama-index-llms-openai",
    "llama-index-llms-openrouter",
    "llama-index-embeddings-nvidia",
    "llama-index-postprocessor-nvidia-rerank",
    "lxml ~= 5.3.0",
    "qdrant-client ~= 1.12.1",
    "torch ~= 2.2.0",
    "transformers ~= 4.42.4",
    "accelerate ~= 0.27.0",
]

[project.optional-dependencies]
dev = [
    "black ~= 24.4",
    "bumpver ~= 2023.1129",
    "flake8 ~= 7.1",
    "isort ~= 5.12",
    "pre-commit ~= 3.5",
]
test = [
    "aiohttp ~= 3.10.5",
    "pytest ~= 7.4",
    "pytest-asyncio ~= 0.21",
]

[tool.bumpver]
current_version = "0.0.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"'
]

[tool.isort]
profile = "black"
src_paths = ["src/ai_pi", "tests"]
combine_as_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["ai_pi"]
concurrency = ["greenlet", "thread"]

[tool.coverage.report]
show_missing = true
omit = ["tests/*"]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["ai_pi"]
